### 2024-02-09 #6342 @editorialbot
[PRE REVIEW]: BiGGer: A Model Transformation Tool written in Java for Bigraph Rewriting in GrGen.NET
**Submitting author:** @PioBeat (<a href="http://orcid.org/0000-0001-6334-2356">Dominik Grzelak</a>)
**Repository:** https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
