### 2024-02-09 #6342 @danielskatz
👋 @PioBeat - thanks for your submission. Before we get started with the review, there are a couple of issues:

1. Your paper is too long. JOSS has a normal upper limit of roughly 1000 words, and your paper is close to 2000.  Please see the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) for an example of a typical JOSS paper.

2. In addition, please work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again.

Once you have made changes in the paper, whether this is content or references, please use the command `@editorialbot generate pdf` to make a new PDF.

editorialbot commands need to be the first entry in a new comment.

I'll put this on pause until you tell me it's ready to start because these two issues have been resolved.

Finally, regarding the other planned submission, since that hasn't yet happen, it doesn't matter to JOSS. But when you do submit that other paper, be sure to either mention or cite this paper, depending on its status, as I don't know the other journal's policy.
