### 2024-02-12 11f83e7
joss: missing DOIs added
