+++ paper.bib
@@ -70,7 +70,8 @@
   pages = {243--258},
   publisher = {{Springer Berlin Heidelberg}},
   isbn = {978-3-540-44618-7},
-  langid = {english}
+  langid = {english},
+  doi = {10.1007/3-540-44618-4_19}
 }
 
 @incollection{milner_EmbeddingsContextsLink_2005,
@@ -161,7 +162,8 @@
   pages = {383--397},
   publisher = {{Springer Berlin Heidelberg}},
   isbn = {978-3-540-38872-2},
-  langid = {english}
+  langid = {english},
+  doi = {10.1007/11841883_27}
 }
 
 @report{jakumeit_GrGenNETUser_2023,
@@ -247,7 +249,8 @@
   pages = {17--33},
   publisher = {{Springer Berlin Heidelberg}},
   isbn = {978-3-662-45917-1},
-  langid = {english}
+  langid = {english},
+  doi = {10.1007/978-3-662-45917-1_2}
 }
 @article{birkedal_MatchingBigraphs_2007,
   title = {Matching of {{Bigraphs}}},
@@ -287,7 +290,8 @@
   location = {{Berlin, Heidelberg}},
   url = {http://dl.acm.org/citation.cfm?id=1812941.1812978},
   urldate = {2018-07-27},
-  isbn = {978-3-642-03740-5}
+  isbn = {978-3-642-03740-5},
+  doi = {10.1007/978-3-642-03741-2_28},
 }
 @inproceedings{sevegnani_BigraphERRewritingAnalysis_2016,
   title = {{{BigraphER}}: {{Rewriting}} and {{Analysis Engine}} for {{Bigraphs}}},
