### 2024-02-12 55634bc
joss: paper length reduced (~1000 words)
