+++ paper.md
@@ -37,10 +37,10 @@ With regard to BiGGer, bigraphs are specified using a meta-modeling approach [@g
 Grounding bigraphs on metamodels facilitate its construction via a universal and platform-agnostic language.
 
 Ultimately, this library transforms EMOF-complaint bigraphs into multigraphs that GrGen.NET can visualize and execute.
-Furthermore, BiGGer is also shipped as a command-line tool for using the functionality also via the terminal for experimentation.
-The most challenging aspect was the accurate translation of bigraphical rules to SPO-based rules, given that bigraphs and GrGen.NET employ distinct approaches regarding graph rewriting (refer to Section "Background").[^3]
+Furthermore, BiGGer is also shipped as a command-line tool for using the functionality via the terminal for experimentation.
+The most challenging aspect was the accurate translation of bigraphical rules to SPO-based rules, given that bigraphs and GrGen.NET employ distinct approaches regarding graph rewriting.[^3]
 
-[^3]: For a comprehensive analysis of the interrelated aspects of these approaches, readers can refer to the works of [@milner_EmbeddingsContextsLink_2005] and [@ehrig_BigraphsMeetDouble_2004].
+[^3]: For an analysis of the interrelated aspects of these approaches, refer to [@milner_EmbeddingsContextsLink_2005; @ehrig_BigraphsMeetDouble_2004].
 
 
 # Statement of Need
@@ -57,21 +57,24 @@ Until now, it is unknown whether the test for graph isomorphism can be solved in
 Subgraph isomorphism is computationally more complex than graph isomorphism and is known to be NP-complete due to the combinatorial nature of finding subpatterns in graphs.
 Thus, graph rewriting is NP-complete [@bacci_FindingForestTree_2014].
 
-However, bigraphs require the matching and rewriting of two substructures; refer to Section "Background".
+However, bigraphs require the matching and rewriting of two substructures.
 Specifically, one substructure of a bigraph is a *forest*, which is relevant to the problem of locating forests in trees [@bacci_FindingForestTree_2014], while the other substructure is a *hypergraph*, which is associated with the subhypergraph matching problem.
-To date, practical bigraph matching primarily exists as a CSP implementation [@miculan_CSPImplementationBigraph_2014] or SAT-based algorithm [@sevegnani_SATBasedAlgorithm_2010] in the literature (refer also to Section "State of the Field").
+To date, practical bigraph matching primarily exists as a CSP implementation [@miculan_CSPImplementationBigraph_2014], or SAT-based algorithm [@sevegnani_SATBasedAlgorithm_2010] in the literature.
 
-However, GrGen.NET provides a universal framework written in C# implementing the so-called Single-Pushout (SPO) approach of graph rewriting for multigraphs (see, for example, [@ehrig_FundamentalsAlgebraicGraph_2006; @jakumeit_GrGenNETUser_2023] and Section "Background").
+GrGen.NET provides a universal framework written in C# implementing the so-called Single-Pushout (SPO) approach of graph rewriting for multigraphs (see, for example, [@ehrig_FundamentalsAlgebraicGraph_2006; @jakumeit_GrGenNETUser_2023]).
 The choice to use GrGen.NET instead of other common graph rewrite systems in the software domain (e.g., AGG, PROGRES, FUJABA) is its execution performance with respect to matching and rewriting (refer to the benchmark described in [@geiss_ImprovementsVarroBenchmark_2007], which was introduced in [@varro_BenchmarkingGraphTransformation_2005] first).
 Specifically, GrGen.NET's pattern matching engine employs some techniques that increase the practical execution performance of subgraph matching, which is an NP-complete problem but is at least one order of magnitude faster than comparable candidates [@jakumeit_GrGenNETUser_2023].
[…]
