### 2024-02-12 95ea575
concurrent append problem: return rule improved (optimized for GrGen.NET)
