### 2024-03-04 #6342 @vissarion
Hi, @danielskatz, thanks for the invitation but my Java experience is very limited, so maybe there is a chance for a more suitable editor.
