### 2024-03-06 #6342 @vissarion
:wave: @maxkfranz, @MiftariB, @hugoledoux, @seanpue, @geekysquirrel, it seems that you are all interested in graph problems and algorithms, would any of you be willing to review this submission for JOSS?
