### 2024-03-08 #6342 @PioBeat
@maxkfranz If that helps you, GrGen.NET can be run under Linux with "monodevelop".
I think I will attempt to provide a Docker image. That would certainly make the review easier.

@vissarion Currently I do not have any reviewer in mind but I will think about that. Thanks for your help.
