### 2024-03-14 #6342 @vissarion
Hi  @gabrielspadon, @sonal-raj, @camillescott, @Materials-Informatics-Laboratory, @idoby would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
