### 2024-03-15 #6342 @idoby
> Hi @gabrielspadon, @sonal-raj, @camillescott, @Materials-Informatics-Laboratory, @idoby would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @vissarion. This isn't my forte but I can try to pretend to say something intelligent during the review. Please consider me as a reviewer but I also advise to try to bring in three reviewers if possible since I'm not confident about my ability to provide an insightful review here.
