### 2024-05-14 #6491 @idoby
Hi, thanks for the reminder!

While this is well outside my field of expertise, just looking at the code it seems that this is a rather thin tool built on top of the GrGen framework. Can the authors please explain why this package represents the results of significant scholarly effort as JOSS requires? It seems to me that if this code was written as part of a research paper on bigraphs, it might belong there rather than be published as a stand-alone general-use tool or library.
Furthermore, the "statement of need" section of the paper does not seem to present the need for this tool, rather it describes what it accomplishes. Hence, I am not convinced that this package should be published as a stand-alone tool.

I would appreciate it if the authors could clarify these points.
