### 2024-05-16 #6491 @PioBeat
Sure, I will be glad to clarify the purpose of this tool!

Note that the present tool paper is part of another journal paper entitled "Efficient Bigraph Rewriting using GrGen.NET" submitted for review to ACM Formal Aspects of Computing, where the full mathematical details are given on how the method actually works that is implemented as presented here for JOSS.
That is, the tool is a usable evaluation in form of a library and command-line tool.
I could provide the manuscript to the reviewers upon request.
While the code size is relatively small, the crucial and difficult part was the translation of the so-called relative pushout (RPO) rewrite semantics of bigraphs to the so-called single pushout (SPO) approach.
These are two different category-theoretic approaches to graph rewriting with different properties.
A side effect of developing such translation was the implementation of so-called tracking rules for bigraphs, which are further useful for causal reasoning and synchronization between two models.

In the software engineering communities it is widely acknowledged that graphs are a appropriate underlying mathematical formalism for representing and analyzing software models, modeling languages, processes and systems.
An approach to model and analyze the dynamics of systems and processes is by using an approach called *graph transformation systems*.
The usefulness of graph transformation in addressing various challenges in general software engineering is extensively explicated in [EEPT06].
Especially the algebraic formalisms, i.e., *algebraic graph transformations systems*, prove valuable as a formal specification technique in distributed state-based systems.
Through the application of so-called *graph transformation rules*, computations and transitions can be effectively described by local transformation of states, which are represented as graphs.
A rule consists of a left-hand side, specifying the graph pattern that needs to be found in the host graph, and a right-hand side that describes how the matched graph elements are rewritten afterwards.
Then, rules are applied to the current graph as long as the left-hand side occurs in the graph.
In this way, a transition system can be created and expanded incrementally by repeating the procedure on all new states until no rule can be applied anymore.

However, graph transformation essentially refers to the subgraph isomorphism problem, and more specifically for bigraphs this is referred to finding the forest in a tree (see [BaMR14]), both are NP-complete.

That being said, implementing bigraph matching and rewriting using the transformation tool GrGen.NET is for the following reasons.
First, to supplement the currently sparsely available rewriting implementations for bigraphs by providing a working and efficient alternative, and second, to compare performance across these approaches in future work.
From a practical standpoint, usable bigraph tools are BigraphER [SeCa16] and jLibBig [ChMP22], which solve bigraph matching as a Boolean satisfiability problem (SAT) and a Constraint satisfaction problem (CSP), respectively.
I believe that it is important to provide alternative rewrite implementation grounded on different techniques to make bigraphs available for future experiments and provide access to users with other tools.
A theory holds limited practical value in the absence of tools for its study and experimental evaluation.

I hope this clarifies some of your points. Should you have any further questions, I am more than happy to assist.

Kind regards,
Dominik

[EEPT06]  Ehrig, Hartmut; Ehrig, Karsten; Prange, Ulrike; Taentzer, Gabriele: Fundamentals of Algebraic Graph Transformation, Monographs in Theoretical Computer Science. An EATCS Series. Berlin Heidelberg : Springer-Verlag, 2006 — ISBN 978-3-540-31187-4
[BaMR14]  Bacci, Giorgio; Miculan, Marino; Rizzi, Romeo: Finding a Forest in a Tree. In: Maffei, M.; Tuosto, E. (Hrsg.): Trustworthy Global Computing, Lecture Notes in Computer Science : Springer Berlin Heidelberg, 2014 — ISBN 978-3-662-45917-1
[SeCa16]  Sevegnani, Michele; Calder, Muffy: BigraphER: Rewriting and Analysis Engine for Bigraphs. In: Chaudhuri, S. ; Farzan, A. (Hrsg.): 28th International Conference on Computer Aided Verification. Bd. 9780. Toronto, Canada : Springer International Publishing, 2016 — ISBN 978-3-319-41539-0
[ChMP22]  Chiapperini, Alessio ; Miculan, Marino ; Peressotti, Marco: Computing (optimal) embeddings of directed bigraphs. In: Science of Computer Programming Bd. 221 (2022)
