### 2024-05-17 #6491 @idoby
Your explanation does sound to me like substantial contribution has been made to your field in the other paper, and hope that it gets accepted into the ACM journal.
However, as I understand it, JOSS requires that the *software* itself represent substantial effort and that it is likely to be used and cited *as software*. This package stands as the implementation of the algorithms in your other paper and therefore I believe users of your algorithms/code will likely cite the other paper, not the JOSS paper (as they should).

To conclude, while I applaud the scientific effort in developing the math, algorithms and implementation behind this package, I still have concerns regarding its fit for JOSS.

Since the JOSS guidelines are somewhat vague on this topic (["Co-publication of science, methods, and software"](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software)), I believe the decision should lie with the editor @vissarion.
