### 2024-05-21 #6491 @vissarion
@idoby @PioBeat I have some news regarding the query scope.

I am convinced (and other JOSS editors agree) that in this case the method and the software can be considered are separate contributions. The software seems to address specific software challenges, it contains useful sample data for testing and evaluation and comes with detailed instructions.

Thus, I will remove the query-scope tag and we could continue the review process. @idoby is this OK with you?
