### 2024-05-25 bigraph-toolkit-suite/bigraphs.grgen-bigraphs#1 @abhishektiwari
Generate and publish Javadoc for BiGGer
Generate, publish, and link javadoc for BiGGer

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/6491)
