### 2024-05-25 bigraph-toolkit-suite/bigraphs.grgen-bigraphs#2 @abhishektiwari
Provide single Dockerfile with Mono, GrGen.NET, and BiGGer
To streamline user workflow, provide single Dockerfile with Mono, GrGen.NET, and BiGGer installed ready to run.

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/6491)
