### 2024-05-25 bigraph-toolkit-suite/bigraphs.grgen-bigraphs#3 @abhishektiwari
Automated tests using Github Action runner
Automated tests and use cases validation on every merge using Github Action runner

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/6491)
