### 2024-05-25 #6491 @abhishektiwari
@PioBeat See my initial comments.

Given paper talks about users being able to use BiGGer programmatically in Java via the API, I suggest to publish the API/interface documentation for BiGGer before accepting this paper. I have opened the ticket to generate and publish the [JavaDoc for BiGGer](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/issues/1). It should be relatively a simple task.

### Nonblocking

- Also, the whole setup requires two different Java versions (Java 11 for GrGen.NET and Java 17 for BiGGer) I suggest guiding users to use something like [jEnv](https://www.jenv.be) to streamline their workflow. Or better provide single [Dockerfile](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/issues/2) with Mono, GrGen.NET, and BiGGer installed and ready to go.

- I can see there test cases (including samples) included with the package but they are not automated. I also recommend to [automate the test using](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/issues/3) the Github action runner.
