### 2024-05-25 bigraph-toolkit-suite/bigraphs.grgen-bigraphs#4 @abhishektiwari
Add contribution/community guidelines
Add contribution/community guidelines [for repository contributors](https://docs.github.com/articles/setting-guidelines-for-repository-contributors).

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/6491)
