### 2024-05-27 4e45e35
minor spelling/formatting corrections
