+++ README.md
@@ -47,14 +47,14 @@ The tool implements a unidirectional transformation from bigraphs to GrGen.NET m
 
 ## Getting Started
 
-Before using the tool or library you have to build the project, or use the pre-compiled JAR, which includes all
+Before using the tool or library, you have to build the project, or use the pre-compiled JAR, which includes all
 dependencies (refer also to [Development](#Development)).
 
 ### Requirements
 
 **GrGen.NET**
 - Mono for execution of GrGen.NET
-  - See on how to install Mono on macOS, Linux or Windows: https://www.mono-project.com/download/stable/#download-lin
+  - See the following link on how to install Mono on macOS, Linux or Windows: https://www.mono-project.com/download/stable/#download-lin
 - Java 11
   - For executing GrGen.NET's visualization (is done via yComp)
   - See also [Troubleshooting](#Troubleshooting)
@@ -206,17 +206,18 @@ The format of this map is in JSON and looks like this:
 }
 ```
 
-Each array represents the element mappings for a specific rule, thus, it must be assigned to a JSON
+Each array represents the element mappings for a specific rule; thus, it must be assigned to a JSON
 object (`"rule1": {"map": [ ... the mapping ... ]}`).
 Either use the given name of the rule or use the generic autogenerated label when none is passed.
-Note that for the latter the order of arguments play an important role when supplying rules without names.
-The array elements indicate the mapping of graph elements from the reactum to the redex of a rule (`[r1,l1]`
-means `r1 -> l1`).
+They follow a simple pattern as depicted above in the example.
+Note that for the latter, the order of arguments plays an important role when supplying rules without names.
+
+The array elements indicate the mapping of graph elements from the reactum to the redex of a rule (`[r1,l1]` means `r1 -> l1`).
 Note that `li` can also be empty (i.e., node addition).
 A graph element can be a node or a link (edge or outer name).
 Additionally, each mapped link has to be explicitly specified under the object `"links"` in each rule JSON object.
 
-This file must contain all tracking maps for all rules.
+This file must contain _all_ tracking maps for _all_ rules.
 
 Example: `--tracking=sample/petrinet-simple/map.json`
 
@@ -225,7 +226,7 @@ Example: `--tracking=sample/petrinet-simple/map.json`
 This option is _optional_.
 
 It specifies the base path for all (meta-)models (tracking map, rules, signature, host graph, ...).
-If not specified, an absolute path or relative path for each model has to be specified.
+If not specified, an absolute path or relative path for each model has to be specified for the respective argument.
 
 ### Output of the Tool
 
@@ -234,7 +235,7 @@ Executing the command creates several files:
 - a metamodel graph `*.gm`
 - the declaration of the initial host graph `*.grs` conforming to the metamodel graph `*.gm`
 - a ruleset file `*.grg` containing all rule specifications
-- a script file named `script.grs` for applying the rules (a simple control strategy is used)
+- a script file named `script.grs` for applying the rules (a simple rule execution strategy is used)
 
 To customize the rule strategy, edit the generated script file `script.grs`.
 
@@ -286,7 +287,7 @@ The following transformers are available:
 - `RuleTransformer`
 Use the concrete implementations of these Java interfaces.
 
-This project includes several demo providers that show how to create signatures, bigraphs and reactules with
+This project includes several demo providers that show how to create signatures, bigraphs and reaction rules with
 Bigraph Framework.
 To model bigraphs, the Java framework called [Bigraph Framework](https://bigraphs.org/products/bigraph-framework/) is
 used.
[…]
