### 2024-05-27 ec89ba5
created contribution/community guidelines and code of conduct (fixes issue #4)
