+++ CONTRIBUTING.md
@@ -1,0 +1,118 @@
+<!-- omit in toc -->
+# Contributing to bigraphs.grgen-bigraphs
+
+All types of contributions are encouraged and valued. See the [Table of Contents](#table-of-contents) for different ways to help and details about how this project handles them. Please make sure to read the relevant section before making your contribution. It will make it a lot easier for us maintainers and smooth out the experience for all involved.
+
+> If you like the project, but just don't have time to contribute, there are other easy ways to support the project and show your appreciation, which we would also be thrilled about:
+> - Star the project
+> - Refer this project in your project's readme
+> - Tell your friends/colleagues
+
+<!-- omit in toc -->
+
+## Table of Contents
+
+- [Code of Conduct](#code-of-conduct)
+- [I Have a Question](#i-have-a-question)
+- [I Want To Contribute](#i-want-to-contribute)
+  - [Reporting Bugs](#reporting-bugs)
+  - [Suggesting Enhancements](#suggesting-enhancements)
+  - [Improving The Documentation](#improving-the-documentation)
+- [Styleguides](#styleguides)
+  - [Commit Messages](#commit-messages)
+- [Join The Project Team](#join-the-project-team)
+
+
+## Code of Conduct
+
+This project and everyone participating in it is governed by the
+[bigraphs.grgen-bigraphs Code of Conduct](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/blob/master/CODE_OF_CONDUCT.md).
+By participating, you are expected to uphold this code. Please report unacceptable behavior
+to <dominik.grzelak@web.de>.
+
+
+## I Have a Question
+
+> If you want to ask a question, we assume that you have read the available [Documentation](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/blob/main/README.md).
+
+Before you ask a question, it is best to search for existing [Issues](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs//issues) that might help you. In case you have found a suitable issue and still need clarification, you can write your question in this issue. It is also advisable to search the internet for answers first.
+
+If you then still feel the need to ask a question and need clarification, we recommend the following:
+
+- Open an [Issue](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs//issues/new).
+- Provide as much context as you can about what you're running into.
+- Provide project and platform versions (Java, Maven, GrGen.NET, etc), depending on what seems relevant.
+
+We will then take care of the issue as soon as possible.
+
+## I Want To Contribute
+
+> ### Legal Notice <!-- omit in toc -->
+> When contributing to this project, you must agree that you have authored 100% of the content, that you have the necessary rights to the content and that the content you contribute may be provided under the project license.
+
+### Reporting Bugs
+
+<!-- omit in toc -->
+#### Before Submitting a Bug Report
+
+A good bug report shouldn't leave others needing to chase you up for more information. Therefore, we ask you to investigate carefully, collect information and describe the issue in detail in your report. Please complete the following steps in advance to help us fix any potential bug as fast as possible.
+
+- Make sure that you are using the latest version.
+- Determine if your bug is really a bug and not an error on your side, e.g. using incompatible environment components/versions (Make sure that you have read the [documentation](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/blob/main/README.md)).
+- To see if other users have experienced (and potentially already solved) the same issue you are having, check if there is not already a bug report existing for your bug or error in the [bug tracker](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/issues?q=label%3Abug).
+- Also make sure to search the internet (including Stack Overflow) to see if users outside the GitHub community have discussed the issue.
+- Collect information about the bug:
[…]
