### 2024-05-28 7e5cadc
added build instruction for generating Java API documentation (now also publically available); modified Readme and community contribution guidelines to point to documentation (solves issue #1)
