+++ CONTRIBUTING.md
@@ -30,7 +30,7 @@ to <dominik.grzelak@web.de>.
 
 ## I Have a Question
 
-> If you want to ask a question, we assume that you have read the available [Documentation](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/blob/main/README.md).
+> If you want to ask a question, we assume that you have read the available [Project Documentation](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/blob/main/README.md) and [Java API Documentation](https://docshoster.org/p/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/latest/introduction.html).
 
 Before you ask a question, it is best to search for existing [Issues](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs//issues) that might help you. In case you have found a suitable issue and still need clarification, you can write your question in this issue. It is also advisable to search the internet for answers first.
 
@@ -55,7 +55,7 @@ We will then take care of the issue as soon as possible.
 A good bug report shouldn't leave others needing to chase you up for more information. Therefore, we ask you to investigate carefully, collect information and describe the issue in detail in your report. Please complete the following steps in advance to help us fix any potential bug as fast as possible.
 
 - Make sure that you are using the latest version.
-- Determine if your bug is really a bug and not an error on your side, e.g. using incompatible environment components/versions (Make sure that you have read the [documentation](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/blob/main/README.md)).
+- Determine if your bug is really a bug and not an error on your side, e.g. using incompatible environment components/versions (Make sure that you have read the [Project Documentation](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/blob/main/README.md) and [Java API Documentation](https://docshoster.org/p/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/latest/introduction.html)).
 - To see if other users have experienced (and potentially already solved) the same issue you are having, check if there is not already a bug report existing for your bug or error in the [bug tracker](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/issues?q=label%3Abug).
 - Also make sure to search the internet (including Stack Overflow) to see if users outside the GitHub community have discussed the issue.
 - Collect information about the bug:
@@ -92,7 +92,7 @@ This section guides you through submitting an enhancement suggestion for bigraph
 #### Before Submitting an Enhancement
 
 - Make sure that you are using the latest version.
-- Read the [documentation](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/blob/main/README.md) carefully and find out if the functionality is already covered, maybe by an individual configuration.
+- Read the [Project Documentation](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/blob/main/README.md) and [Java API Documentation](https://docshoster.org/p/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/latest/introduction.html) carefully and find out if the functionality is already covered, maybe by an individual configuration.
 - Perform a [search](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs//issues) to see if the enhancement has already been suggested. If it has, add a comment to the existing issue instead of opening a new one.
 - Find out whether your idea fits with the scope and aims of the project. It's up to you to make a strong case to convince the project's developers of the merits of this feature. Keep in mind that we want features that will be useful to the majority of our users and not just a small subset. If you're just targeting a minority of users, consider writing an add-on/plugin library.
 
