### 2024-05-28 9a18b87
added GitHub Actions workflow for Maven tests (issue #3)
