### 2024-05-28 5bfb540
added GitHub Actions workflow for Maven tests (issue #3)
