+++ .github/workflows/ci-maven-test.yml
@@ -24,6 +24,9 @@ jobs:
           java-version: ${{ matrix.java-version }}
           distribution: 'temurin'
 
+      - name: Install Graphviz
+        run: sudo apt-get install -y graphviz
+
       - name: Cache local Maven repository
         uses: actions/cache@v4
         with:
