+++ .github/workflows/ci-maven-test.yml
@@ -40,3 +40,14 @@ jobs:
 
       - name: Run tests
         run: mvn test
+
+      - name: Zip test results
+        run: |
+          mkdir -p artifacts
+          zip -r artifacts/test-results.zip src/test/resources/dump
+
+      - name: Upload test results
+        uses: actions/upload-artifact@v4
+        with:
+          name: test-results
+          path: artifacts/test-results.zip
