+++ .github/workflows/ci-maven-test.yml
@@ -45,8 +45,9 @@ jobs:
         run: |
           mkdir -p artifacts
           pwd
-          ls
-          zip -r artifacts/test-results.zip src/test/resources/dump/*
+          ls ./src/test/resources/dump
+          ls ./src/test/resources/dump/
+          zip -r artifacts/test-results.zip ./src/test/resources/dump
 
       - name: Upload test results
         uses: actions/upload-artifact@v4
