+++ .github/workflows/ci-maven-test.yml
@@ -2,9 +2,9 @@ name: Java Maven CI for Tests
 
 on:
   push:
-    branches: [ main, testing ]
+    branches: [ main ]
   pull_request:
-    branches: [ main, testing ]
+    branches: [ main ]
 
 jobs:
   build:
