+++ README.md
@@ -1,13 +1,12 @@
 # BiGGer: A Model Transformation Tool for Bigraph Rewriting with GrGen.NET
 
-**Stable Version:** 1.1.0
-**Development Version:** 2.0-SNAPSHOT
+**Stable Version:** 1.1.0 <a href='https://docshoster.org/p/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/1.1.0/introduction.html'><img src='https://docshoster.org/pstatic/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/1.1.0/badge.svg'/>
+</a>
+
+**Development Version (Latest):** 2.0-SNAPSHOT <a href='https://docshoster.org/p/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/2.0-SNAPSHOT/introduction.html'><img src='https://docshoster.org/pstatic/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/2.0-SNAPSHOT/badge.svg'/>
+</a>
 
 [![DOI](https://zenodo.org/badge/715108447.svg)](https://zenodo.org/doi/10.5281/zenodo.10803175)
-
-<a href='https://docshoster.org/p/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/latest/introduction.html'>
-<img src='https://docshoster.org/pstatic/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/latest/badge.svg'/>
-</a>
 
 ----
 
@@ -38,7 +37,7 @@ The tool implements a unidirectional transformation from bigraphs to GrGen.NET m
 - Place-Sorts
 - Link-sorts
 - Rule control instructions (however, a default one is provided and can be customized)
-- Instantiation map of a rule
+- Instantiation map of a reaction rule
 - User-defined attributes on bigraph nodes/edges
 
 **Additional Features**
@@ -52,8 +51,7 @@ The tool implements a unidirectional transformation from bigraphs to GrGen.NET m
 
 ## Getting Started
 
-Before using the tool or library, you have to build the project, or use the pre-compiled JAR, which includes all
-dependencies (refer also to [Development](#Development)).
+Before using the tool or library, you have to build the project (refer to [Development](#Development)), or use the shipped JAR, which is available via [GitHub Releases](https://github.com/bigraph-toolkit-suite/bigraphs.grgen-bigraphs/releases).
 
 ### Requirements
 
@@ -199,7 +197,7 @@ Example: `--rule=rule3:sample/petrinet-simple/r3-lhs.xmi,sample/petrinet-simple/
 This option is _optional_.
 But when **--rule** ist used, this option is _mandatory_.
 
-Additionally, for each rule a tracking map has to be supplied.
+Additionally, for each rule a "tracking map" has to be supplied.
 The format of this map is in JSON and looks like this:
 
 ```
@@ -217,8 +215,8 @@ Either use the given name of the rule or use the generic autogenerated label whe
 They follow a simple pattern as depicted above in the example.
 Note that for the latter, the order of arguments plays an important role when supplying rules without names.
 
-The array elements indicate the mapping of graph elements from the reactum to the redex of a rule (`[r1,l1]` means `r1 -> l1`).
-Note that `li` can also be empty (i.e., node addition).
+The array elements indicate the mapping of graph element IDs from the reactum to the redex of a rule (`[R1,L1]` means `R1 -> L1`).
+Note that `Li` can also be empty (i.e., node addition).
 A graph element can be a node or a link (edge or outer name).
 Additionally, each mapped link has to be explicitly specified under the object `"links"` in each rule JSON object.
 
@@ -280,7 +278,7 @@ trackingMap.addLinkNames("e1");
 t.withMap(trackingMap);
[…]
