### 2024-05-29 #6491 @PioBeat
Dear @abhishektiwari,

- [x] I have created the required contribution/community guidelines (CONTRIBUTING and CODE_OF_CONDUCT).

- [x] I have uploaded the Javadoc for the library and linked it in the README and community guidelines.

- [x] I added the Github Action for automating tests and validating the use cases.

I plan to create the Dockerfile with Mono, GrGen.NET, and BiGGer next week.
Thank you for pointing out jEnv. This seems also a viable option.
