### 2024-06-03 d5adc1a
Provide single Dockerfile with Mono, GrGen.NET, and BiGGer (solves #2)
