### 2024-06-03 #6491 @PioBeat
I have added a Dockerfile with Mono, GrGen.NET, and BiGGer already installed.
The README explains how to run this.

I will close the respective issues after some time if there are no further amendments from your side.
