### 2024-06-11 #6491 @vissarion
Thanks @PioBeat . Please update your zenodo record such that the title/authors/affiliations are exactly the same as in the JOSS submission.
