### 2024-06-11 bigraph-toolkit-suite/bigraphs.grgen-bigraphs#3 @PioBeat
implemented in v1.2.0
