### 2024-06-11 bigraph-toolkit-suite/bigraphs.grgen-bigraphs#1 @PioBeat
implemented in v1.2.0
