### 2024-06-11 553c6c4
small changes in paper.md
