### 2024-06-11 bd0d6c6
case changes in paper.bib
