+++ paper.bib
@@ -11,7 +11,7 @@
 }
 
 @book{steinberg_EMFEclipseModeling_2008,
-  title = {{{EMF}}: {{Eclipse Modeling Framework}}},
+  title = {{EMF}: Eclipse Modeling Framework},
   shorttitle = {{{EMF}}},
   author = {Steinberg, Dave and Budinsky, Frank and Paternostro, Marcelo},
   date = {2008-12-16},
@@ -39,7 +39,7 @@
 }
 
 @book{milner_SpaceMotionCommunicating_2009,
-  title = {The {{Space}} and {{Motion}} of {{Communicating Agents}}},
+  title = {The Space and Motion of Communicating Agents},
   author = {Milner, Robin},
   date = {2009},
   edition = {1},
@@ -75,12 +75,12 @@
 }
 
 @incollection{milner_EmbeddingsContextsLink_2005,
-  title = {Embeddings and {{Contexts}} for {{Link Graphs}}},
-  booktitle = {Formal {{Methods}} in {{Software}} and {{Systems Modeling}}: {{Essays Dedicated}} to {{Hartmut Ehrig}} on the {{Occasion}} of {{His}} 60th {{Birthday}}},
+  title = {Embeddings and Contexts for Link Graphs},
+  booktitle = {Formal Methods in Software and Systems Modeling: Essays Dedicated to {{Hartmut Ehrig}} on the Occasion of His 60th Birthday},
   author = {Milner, Robin},
   editor = {Kreowski, Hans-Jörg and Montanari, Ugo and Orejas, Fernando and Rozenberg, Grzegorz and Taentzer, Gabriele},
   date = {2005},
-  series = {Lecture {{Notes}} in {{Computer Science}}},
+  series = {Lecture Notes in Computer Science},
   pages = {343--351},
   publisher = {{Springer}},
   location = {{Berlin, Heidelberg}},
@@ -91,21 +91,21 @@
 
 @incollection{ehrig_BigraphsMeetDouble_2004,
   title = {Bigraphs Meet Double Pushouts},
-  booktitle = {Current {{Trends}} in {{Theoretical Computer Science}}},
+  booktitle = {Current Trends in Theoretical Computer Science},
   author = {Ehrig, Hartmut},
   date = {2004-04-01},
   pages = {27--40},
-  publisher = {{WORLD SCIENTIFIC}},
+  publisher = {World Scientific},
   doi = {10.1142/9789812562494_0038},
   isbn = {978-981-238-783-7},
   volumes = {0}
 }
 
 @book{ehrig_FundamentalsAlgebraicGraph_2006,
-  title = {Fundamentals of {{Algebraic Graph Transformation}}},
+  title = {Fundamentals of Algebraic Graph Transformation},
   author = {Ehrig, Hartmut and Ehrig, Karsten and Prange, Ulrike and Taentzer, Gabriele},
   date = {2006},
-  series = {Monographs in {{Theoretical Computer Science}}. {{An EATCS Series}}},
+  series = {Monographs in Theoretical Computer Science. {A}n {EATCS} Series},
   publisher = {{Springer-Verlag}},
   location = {{Berlin Heidelberg}},
   doi = {10.1007/3-540-31188-2},
@@ -114,12 +114,12 @@
 }
 
 @inproceedings{chiapperini_ComputingEmbeddingsDirected_2020,
-  title = {Computing {{Embeddings}} of {{Directed Bigraphs}}},
-  booktitle = {Graph {{Transformation}}},
+  title = {Computing Embeddings of Directed Bigraphs},
+  booktitle = {Graph Transformation},
   author = {Chiapperini, Alessio and Miculan, Marino and Peressotti, Marco},
[…]
