### 2024-06-11 #6491 @danielskatz
@PioBeat - please take a look at this proof and confirm it looks ok. I will also look at it.
